update: Seperate ProjectService -> ProjectService and PermissionService
