+++ src/main/java/com/example/demo/repository/PermissionRepository.java
@@ -1,0 +1,8 @@
+package com.example.demo.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.security.Permission;
+
+public interface PermissionRepository extends JpaRepository<Permission, Long> {
+}
